requirements.txt
python-telegram-bot==22.7
requests==2.32.5
beautifulsoup4==4.14.3
soupsieve==2.8.3
python-dotenv==1.2.2
boto3
